To check the efficiency of the Smart's Algorithm in SageMath
This was created to check the efficiency of the Smart's Algorithm in SageMath, it uses the time package to calculate the total time taken to run this attack algorithm which is then plotted on a graph using SageMath. The alogithm is adjusted sightly to make sure we are only concerned with evaluating the efficient time. We work until prime numbers with a 24-bit length which is due to limitations in finding Anomalous Curves with a laptop. Trying to find Anomalous curves of order p where p is a 8 digit number takes more than 6 hours on my current system (Arch Linux, Intel® Core™ i7-7500U, 8GB RAM) and has not given me a curve because it hangs the system after 6 hours of running to find a single Anomalous curve of a 8 digit prime p.